flask-server/requirements.txt
flask
python-dotenv
flask_sqlalchemy
flask-bcrypt
flask-sessions
flask_cors
flask_session
flask_socketio
pymysql
apscheduler
pandas
numpy
simpy
matplotlib
openpyxl
IPython
twilio
pymongo
argon2-cffi
